Safari-609.1.11.1/LayoutTests/platform/ios-simulator/webgl/2.0.0/conformance2/textures/misc/tex-3d-size-limit-expected.txt

This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../../resources/webgl_test_files/conformance2/textures/misc/tex-3d-size-limit.html
[ 1: PASS ] gl = wtu.create3DContext(undefined, undefined, 2) is non-null.
[ 2: FAIL ] getError expected: INVALID_VALUE. Was INVALID_ENUM : texImage3D should fail for dimension out of range.
[ 3: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 4: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 5: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 6: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 7: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 8: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 9: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 10: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 11: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 12: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 13: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 14: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 15: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 16: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 17: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 18: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 19: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 20: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 21: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 22: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 23: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 24: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 25: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 26: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 27: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 28: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 29: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 30: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 31: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 32: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 33: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 34: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 35: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 36: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 37: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 38: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 39: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 40: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 41: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 42: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 43: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 44: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 45: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 46: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 47: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 48: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 49: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 50: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 51: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 52: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 53: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 54: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 55: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 56: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 57: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 58: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 59: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 60: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 61: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 62: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 63: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 64: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 65: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 66: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 67: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 68: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 69: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 70: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 71: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 72: PASS ] getError was expected value: NO_ERROR : texImage3D should succeed.
[ 73: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 74: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 75: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : texImage3D should fail for dimension out of range.
[ 76: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : ArrayBufferView not big enough for request by texImage3D.
[ 77: PASS ] getError was expected value: NO_ERROR : ArrayBufferView big enough for request by texImage3D.
[ 78: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : ArrayBufferView not big enough for request by texSubImage3D.
[ 79: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : ArrayBufferView not big enough for request by texImage3D.
[ 80: PASS ] getError was expected value: NO_ERROR : ArrayBufferView big enough for request by texImage3D.
[ 81: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : ArrayBufferView not big enough for request by texSubImage3D.
[ 82: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : ArrayBufferView not big enough for request by texImage3D.
[ 83: PASS ] getError was expected value: NO_ERROR : ArrayBufferView big enough for request by texImage3D.
[ 84: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : ArrayBufferView not big enough for request by texSubImage3D.
[ 85: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : ArrayBufferView not big enough for request by texImage3D.
[ 86: PASS ] getError was expected value: NO_ERROR : ArrayBufferView big enough for request by texImage3D.
[ 87: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : ArrayBufferView not big enough for request by texSubImage3D.
[ 88: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : ArrayBufferView not big enough for request by texImage3D.
[ 89: PASS ] getError was expected value: NO_ERROR : ArrayBufferView big enough for request by texImage3D.
[ 90: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : ArrayBufferView not big enough for request by texSubImage3D.
[ 91: PASS ] successfullyParsed is true
[ FAIL ] 48 failures reported
